Initial upload of working Jupyter Notebooks
The D&C Webscraping Starter was my initial notebook, in which I developed my code for webscraping information about the scriptures I needed from the website of the Church of Jesus Christ of Latter-day Saints.

After getting that settled, I used the Book of Mormon Webscraping File notebook to refine the code, and to get the information I need about all the standard works scraped and downloaded to my device.