Finished markdowns, began EDA
I finished markdown text explaining what was going on throughout the notebook, and made sure to annotate the operations of each cell that had not been. Basically, I cleaned this notebook up to be more presentable and digestible to myself and to any other future reader.

I also began working on some EDA of the data I collected. I'll continue that EDA on Monday to refresh myself on how to do all that in Python, and then I'll move on to cleaning it better and doing more webscraping on Tuesday.