Created General Conference Webscraper notebook
I realized that with the approaching General Conference, I had a more pressing breakdown maker that I needed to make. This one should be (more or less) adaptable and adjustable to fit my needs for the Come, Follow Me Breakdown. If it isn't, it will at least provide me with a solid base upon which to build that tool that this project is really all about.