Created Regex Learning notebook
This notebook was all about learning the basics of Regular Expressions, which I'll need to make use of as I do my webscraping for time durations of conference talks and which I'll need for getting, at the very least, the start and end dates for the Come, Follow Me breakdown.

I think. This may have just been a pleasant, though potentially ultimately useless, detour.