Progressed through learning about dates
One final thing I'll need to learn about is working with dates. During my bootcamp, there was a "dealing with dates" module that was optional, and given that the mandatory coursework was so demanding, I never had time to get to that particular optional module. However, since I'll be working with dates in both my Conference Calculator and my Come, Follow Me Calculator, I figure now is as good a time as any to establish a solid foundation of understanding and working with dates so I can work with them competently."