Renamed Conference Calculator, copied to an improvement notebook
I renamed the Final Conference Calculator to Conference Calculator 1.0. This renaming resulted in the "deletion" of the originally named notebook.

I also copied that notebook to make the Conference Calculator Improvement notebook, so I can keep the current calculator "live" and not mess it up while I work on making improvements to it.