Updated Line Counter Calculator and Directory Locations
On further consideration, I decided that I wanted to count the lines as closely to printed scripture as possible, so I had to spend time in the D&C Webscraping starter notebook figuring out the best way to do that. I decided on artificially inflating the number of lines counted to the appropriate level by multiplying the line height by 0.67, which causes the number of lines calculated to fit inside the created rectangle of each verse to be elevated. This level of adjustment produced the results most accurate to printed scriptures.

I then applied that adjustment to the 'get_verse_lines()' function of the Book of Mormon Webscraping notebook.

I also went through both notebooks and updated directory locations since I had renamed a few folders.