began adjusting organization of readings_df
readings_df is the dataframe that contains the information about each week's readings. I started adjusting that dataframe so that each row contains the group number, group assignment, and the assignments and links for each assignment for each week in that group number.

I plan to keep working this to whittle that dataframe down to just having a single row for each group assignment.