Use regex to get chapter columns
Not a lot of work put in today, but it was something. I just used regex to get the names of the chapters.

Tomorrow (8 Nov) I need to iterate through the chapters_cols and through each row of the readings_df. I need to then split the columns on the '–' **only** if they do not have a ':' (so as to not cause problems for the Joseph Smith—History assignments), establish one column as the starting chapter, create a new column for every chapter between the starting and ending chapter, then establish the ending chapter as the end chapter column.

I'll also need to make sure, at some point before getting too far ahead, that this will work for assignments that span multiple books - I don't know why it wouldn't, but I need to be on the lookout!