Get Video, Paragraph, and Line Length and Role of Speaker
I finally modularized the section of the code that gets the length of each talk in time, paragraphs, and lines, which necessarily relied on finding why the play button wasn't being found.

I also made one of the main adjustments to the code that I wanted to see in the calculator 2.0: making sure that picture captions and descriptions are not counted or included in the length of the talk.