Finished all aspects of calculator related to webscraping
I set up the calculator to get the role of the author earlier in the process, and I condensed the user input to the front of the program.

I also set up the calculator to scrape and export the text of the talks in the same function in which it scrapes information about the length of the talks. The overall effect of this is to reduce the time spent webscraping and calculating information about each of the talks.