finished modularizing components of the calculator
That's it. I finished modularizing the 8 different aspects of the calculator, and it all works great, exactly like I hoped it would.

The end result is a much faster program that will be easier to update and fix and mess around with in the future.

I'm happy with and proud of that, and I'm excited to have a job that keeps me happy and busy enough to not have much time to work on this.